framework/persistence/src/main/java/org/koenighotze/jee7hotel/framework/persistence/converter/LocalDateTimeConverter.java
package org.koenighotze.jee7hotel.framework.persistence.converter;

import javax.persistence.AttributeConverter;
import javax.persistence.Converter;
import java.sql.Timestamp;
import java.time.LocalDateTime;

/**
 * @author dschmitz
 */
@Converter(autoApply = true)
public class LocalDateTimeConverter implements AttributeConverter<LocalDateTime, Timestamp> {
    @Override
    public Timestamp convertToDatabaseColumn(LocalDateTime localDateTime) {
        if (null == localDateTime) {
            return null;
        }
        return Timestamp.valueOf(localDateTime);
    }

    @Override
    public LocalDateTime convertToEntityAttribute(Timestamp dbData) {
        if (null == dbData) {
            return null;
        }
        return LocalDateTime.from(dbData.toInstant());
    }
}
